refactored

+++ food_kart/src/main/java/org/food_kart/entity/Review.java
@@ -2,12 +2,14 @@ package org.food_kart.entity;
 
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 public class Review {
     private int reviewId;
     private String restaurantName;
     private int restaurantId;
-    private int rating;
+    private double rating;
     private String comment;
     
 }
